🚚 Rename: FolderLinkAddBar의 이름을 HTML태그로서 역할을 명확히 함
